Extract command palette actions creation to separate module
