+++ b4n/ui/views/resources/view.rs
@@ -2,13 +2,11 @@ use b4n_common::NotificationSink;
 use b4n_config::PluginRef;
 use b4n_config::keys::KeyCommand;
 use b4n_kube::plugins::PluginContext;
-use b4n_kube::{
-    ALL_NAMESPACES, CONTAINERS, EVENTS, Kind, NAMESPACES, NODES, Namespace, ObserverResult, PODS, Port, ResourceRef, SECRETS,
-};
+use b4n_kube::{CONTAINERS, EVENTS, Kind, NODES, Namespace, ObserverResult, PODS, Port, ResourceRef};
 use b4n_list::Row;
 use b4n_tui::table::{Table, ViewType};
-use b4n_tui::widgets::{ActionItem, ActionsList, ActionsListBuilder, Button, CheckBox, Dialog, Selector, ValidatorKind};
-use b4n_tui::{MouseEventKind, PluginsExt, ResponseEvent, Responsive, ScopeData, ToSelectData, TuiEvent};
+use b4n_tui::widgets::{ActionsList, ActionsListBuilder, Button, CheckBox, Dialog, Selector};
+use b4n_tui::{MouseEventKind, ResponseEvent, Responsive, ScopeData, ToSelectData, TuiEvent};
 use delegate::delegate;
 use kube::{config::NamedContext, discovery::Scope};
 use ratatui::layout::Position;
@@ -17,11 +15,13 @@ use std::{collections::HashMap, path::PathBuf, rc::Rc};
 
 use crate::core::{PreviousData, ResourcesInfo, SharedAppData, SharedAppDataExt, SharedBgWorker};
 use crate::kube::extensions::ActionsListBuilderExt;
-use crate::kube::resources::pod::PF_COLUMN_NO;
 use crate::kube::resources::{ResourceItem, ResourcesList, node, pod};
 use crate::ui::views::View;
+use crate::ui::views::resources::menu::{
+    build_create_resource_actions, build_mouse_menu_actions, build_port_forward_steps, build_resources_actions,
+};
 use crate::ui::views::resources::{NextRefreshActions, table::ResourcesTable};
-use crate::ui::widgets::{CommandPalette, Filter, NamespaceSelector, StepBuilder};
+use crate::ui::widgets::{CommandPalette, Filter, NamespaceSelector};
 
 /// Resources view (main view) for `b4n`.
 pub struct ResourcesView {
@@ -217,30 +217,8 @@ impl ResourcesView {
     /// Displays a list of available forward ports for a container to choose from.
     pub fn show_ports_list(&mut self, list: &[Port]) {
         if let Some(resource) = self.table.get_resource_ref(true) {
-            let actions_list = ActionsListBuilder::from_resource_ports(list).build(None);
-            self.command_palette = CommandPalette::new(Rc::clone(&self.app_data), actions_list, 65)
-                .with_header(format!(
-                    " Add port forward for '{}' container:",
-                    resource.container.as_deref().unwrap_or_default()
-                ))
-                .with_prompt("container port")
-                .with_validator(ValidatorKind::Number(0, 65_535))
-                .with_step(
-                    StepBuilder::input("")
-                        .with_validator(ValidatorKind::Number(0, 65_535))
-                        .with_prompt("local port")
-                        .with_colors(self.app_data.borrow().theme.colors.command_palette.clone())
-                        .build(&self.app_data),
-                )
-                .with_step(
-                    StepBuilder::input("127.0.0.1")
-                        .with_validator(ValidatorKind::IpAddr)
-                        .with_prompt("bind address")
-                        .with_colors(self.app_data.borrow().theme.colors.command_palette.clone())
-                        .build(&self.app_data),
-                )
-                .with_response(|v| build_port_forward_response(v, resource))
-                .with_highlighted_position(self.last_mouse_click.take());
+            self.command_palette =
+                build_port_forward_steps(&self.app_data, resource, list).with_highlighted_position(self.last_mouse_click.take());
             self.command_palette.show();
         }
     }
@@ -371,77 +349,7 @@ impl ResourcesView {
             return;
         }
 
-        let is_selected = self.table.list.table.is_anything_selected();
-        let is_highlighted = self.table.list.table.is_anything_highlighted();
-        let is_containers = self.table.kind_plural() == CONTAINERS;
-        let is_pods = self.table.kind_plural() == PODS;
-        let is_events = self.table.kind_plural() == EVENTS;
-        let is_deletable = is_selected && self.table.list.table.data.is_deletable;
-
-        let mut builder = ActionsListBuilder::from_kinds(self.app_data.borrow().kinds.as_deref())
-            .with_resources_actions(!is_containers && is_deletable)
-            .with_forwards()
-            .with_filter_action("filter")
-            .with_pin_filter_action("pin_filter")
-            .with_actions(
-                self.app_data
-                    .borrow()
-                    .plugins
-                    .to_actions(self.table.get_kind().as_str(), is_highlighted, is_selected),
-            );
-
-        if self.table.kind_plural() != NAMESPACES {
-            builder.add_action(
-                ActionItem::action("back", "back").with_description("returns to the previous view"),
-                Some(KeyCommand::NavigateBack),
-            );
-        }
-
-        if self.table.scope() == &Scope::Namespaced && !is_containers {
-            builder = builder.with_namespace();
-        }
-
-        let selected = if is_selected { "selected" } else { "all" };
-        builder.add_action(
-            ActionItem::action("copy", "copy").with_description(&format!("copies {selected} resources to clipboard")),
-            Some(KeyCommand::ContentCopy),
-        );
-
-        if !is_containers && !is_events {
-            if is_highlighted {
-                builder.add_action(
-                    ActionItem::action("show events", "show_events").with_description("shows events for the selected resource"),
-                    Some(KeyCommand::EventsShow),
-                );
-            }
-
-            if self.table.list.table.data.is_creatable {
-                builder.add_action(
-                    ActionItem::action("create", "create")
-                        .with_description("creates new Kubernetes resource")
-                        .with_aliases(["new", "add"]),
-                    Some(KeyCommand::YamlCreate),
-                );
-            }
-        }
-
-        if self.has_involved_object() {
-            builder.add_action(
-                ActionItem::action("involved object", "show_involved").with_description("navigates to the involved object"),
-                Some(KeyCommand::InvolvedObjectShow),
-            );
-        }
-
-        if is_highlighted {
-            builder = self.add_resource_actions(builder, is_containers);
-            if is_containers || is_pods {
-                builder = Self::add_container_actions(builder);
-            }
-        }
-
-        builder = builder.with_aliases(&self.app_data.borrow().config.aliases);
-        let actions = builder.build(Some(&self.app_data.borrow().key_bindings));
-
+        let actions = build_resources_actions(&self.app_data, &self.table);
         self.open_command_palette(actions);
     }
 
@@ -452,140 +360,15 @@ impl ResourcesView {
         self.footer_tx.hide_hint();
     }
 
-    fn add_resource_actions(&self, mut builder: ActionsListBuilder, is_containers: bool) -> ActionsListBuilder {
-        if self.table.kind_plural() == SECRETS {
-            builder.add_action(
-                ActionItem::action("decode", "decode_yaml").with_description("shows decoded YAML of the highlighted secret"),
-                Some(KeyCommand::YamlDecode),
-            );
-        }
-
-        if self.table.list.table.data.is_editable {
-            builder.add_action(
-                ActionItem::action("edit YAML", "edit_yaml")
-                    .with_description("displays YAML and switches to edit mode")
-                    .with_aliases(["yaml", "yml", "patch"]),
-                Some(KeyCommand::YamlEdit),
-            );
-        }
-
-        builder
-            .with_action(
-                ActionItem::action("show YAML", "show_yaml")
-                    .with_description(if is_containers {
-                        "shows YAML of the container's resource"
-                    } else {
-                        "shows YAML of the highlighted resource"
-                    })
-                    .with_aliases(["yaml", "yml", "view"]),
-                Some(KeyCommand::YamlOpen),
-            )
-            .with_action(
-                ActionItem::action("describe", "describe").with_description("shows resource describe view"),
-                Some(KeyCommand::DescribeOpen),
-            )
-    }
-
-    fn add_container_actions(builder: ActionsListBuilder) -> ActionsListBuilder {
-        builder
-            .with_action(
-                ActionItem::action("show logs", "show_logs")
-                    .with_description("shows container logs")
-                    .with_aliases(["logs"]),
-                Some(KeyCommand::LogsOpen),
-            )
-            .with_action(
-                ActionItem::action("show previous logs", "show_plogs")
-                    .with_description("shows container previous logs")
-                    .with_aliases(["previous"]),
-                Some(KeyCommand::PreviousLogsOpen),
-            )
-            .with_action(
-                ActionItem::action("attach", "attach").with_description("attaches to container main process"),
-                Some(KeyCommand::ContainerAttach),
-            )
-            .with_action(
-                ActionItem::action("shell", "open_shell").with_description("opens container shell"),
-                Some(KeyCommand::ShellOpen),
-            )
-            .with_action(
-                ActionItem::action("forward port", "port_forward")
-                    .with_description("forwards container port")
-                    .with_aliases(["port", "pf"]),
-                Some(KeyCommand::PortForwardsCreate),
-            )
-    }
-
     fn show_mouse_menu(&mut self, x: u16, y: u16) {
         if !self.app_data.borrow().is_connected() {
             return;
         }
 
-        let mut size = 22;
-        let is_selected = self.table.list.table.is_anything_selected();
-        let highlighted_name = self.table.list.table.get_highlighted_item_name();
-        let is_highlighted = highlighted_name.is_some_and(|n| n != ALL_NAMESPACES);
-        let is_containers = self.table.kind_plural() == CONTAINERS;
-        let is_pods = self.table.kind_plural() == PODS;
-        let is_events = self.table.kind_plural() == EVENTS;
+        let actions = build_mouse_menu_actions(&self.table);
+        let width = u16::try_from(actions.max_item_len() + 4).unwrap_or(u16::MAX).max(22);
 
-        let copy = if is_selected { "selected" } else { "all" };
-        let mut builder = ActionsListBuilder::default()
-            .with_menu_action(ActionItem::command_palette())
-            .with_menu_action(ActionItem::menu(11, &format!("󰆏 copy ␝{copy}␝"), "copy"));
-
-        if self.table.kind_plural() != NAMESPACES {
-            builder.add_menu_action(ActionItem::menu(100, "󰕍 back", "back"));
-        }
-
-        if self.table.list.table.is_anything_selected() && self.table.list.table.data.is_deletable {
-            let action = ActionItem::menu(12, " delete ␝selected␝", "").with_response(ResponseEvent::AskDeleteResources);
-            builder.add_menu_action(action);
-        }
-
-        if !is_containers && !is_events {
-            if self.table.list.table.data.is_creatable {
-                builder.add_menu_action(ActionItem::menu(9, "󰐕 create new", "create"));
-            }
-            if is_highlighted {
-                builder.add_menu_action(ActionItem::menu(98, "󰑏 events", "show_events"));
-            }
-        }
-
-        if self.has_involved_object() {
-            builder.add_menu_action(ActionItem::menu(99, "󰑏 involved object", "show_involved"));
-        }
-
-        if is_highlighted {
-            builder = builder
-                .with_menu_action(ActionItem::menu(4, " describe", "describe"))
-                .with_menu_action(ActionItem::menu(1, " YAML", "show_yaml"))
-                .with_menu_action(ActionItem::menu(11, "󰆏 copy ␝name␝", "copy_name"));
-
-            if is_containers || is_pods {
-                builder = builder
-                    .with_menu_action(ActionItem::menu(2, " logs", "show_logs"))
-                    .with_menu_action(ActionItem::menu(3, " logs ␝previous␝", "show_plogs"))
-                    .with_menu_action(ActionItem::menu(6, " attach", "attach"))
-                    .with_menu_action(ActionItem::menu(7, " shell", "open_shell"))
-                    .with_menu_action(ActionItem::menu(8, "󱘖 forward port", "port_forward"));
-
-                if is_pods && self.has_highlighted_item_active_port_forward() {
-                    size = 25;
-                    builder.add_menu_action(ActionItem::menu(8, " stop ␝port forwards␝", "ask_stop_port_forwards"));
-                }
-            }
-
-            if self.table.kind_plural() == SECRETS {
-                builder.add_menu_action(ActionItem::menu(5, " YAML ␝decoded␝", "decode_yaml"));
-            }
-
-            if self.table.list.table.data.is_editable {
-                builder.add_menu_action(ActionItem::menu(10, " edit", "edit_yaml"));
-            }
-        }
-
-        self.command_palette = CommandPalette::new(Rc::clone(&self.app_data), builder.build(None), size).to_mouse_menu();
+        self.command_palette = CommandPalette::new(Rc::clone(&self.app_data), actions, width).to_mouse_menu();
         self.command_palette.show_at((x.saturating_sub(3), y).into());
     }
 
@@ -598,23 +381,7 @@ impl ResourcesView {
             return;
         }
 
-        let mut builder = ActionsListBuilder::default()
-            .with_menu_action(ActionItem::action("full", "new_full").with_description("get all possible fields for the spec"))
-            .with_menu_action(
-                ActionItem::action("minimal", "new_minimal").with_description("get only required fields for the spec"),
-            );
-
-        if let Some(name) = self.table.list.table.get_highlighted_item_name()
-            && (self.kind_plural() != NAMESPACES || name != ALL_NAMESPACES)
-        {
-            builder = builder.with_menu_action(
-                ActionItem::action("duplicate", "new_clone")
-                    .with_description("use the spec of the highlighted resource")
-                    .with_aliases(["clone"]),
-            );
-        }
-
-        let actions = builder.build(None);
+        let actions = build_create_resource_actions(&self.table, self.kind_plural());
         self.command_palette = CommandPalette::new(Rc::clone(&self.app_data), actions, 65)
             .with_prompt("create new resource")
             .with_first_highlighted()
@@ -745,14 +512,6 @@ impl ResourcesView {
         }
     }
 
-    fn has_involved_object(&self) -> bool {
-        self.table
-            .list
-            .table
-            .get_highlighted_resource()
-            .is_some_and(|res| res.involved_object.is_some())
-    }
-
     fn copy_name_to_clipboard(&mut self) {
         if let Some(res) = self.table.list.table.get_highlighted_resource() {
             self.app_data
@@ -775,15 +534,6 @@ impl ResourcesView {
             let new_list = worker.get_port_forward_refs(namespace);
             self.table.list.table.update_port_forwards(&new_list);
         }
-    }
-
-    fn has_highlighted_item_active_port_forward(&self) -> bool {
-        let Some(resource) = self.table.list.table.get_highlighted_resource().and_then(|r| r.data.as_ref()) else {
-            return false;
-        };
-
-        resource.extra_values.len() > PF_COLUMN_NO
-            && resource.extra_values[PF_COLUMN_NO].raw_text().is_some_and(|t| !t.is_empty())
     }
 
     fn stop_port_forwards(&self) -> ResponseEvent {
@@ -936,17 +686,6 @@ impl View for ResourcesView {
         self.command_palette.draw(frame, area);
         self.filter.draw(frame, area);
         self.namespace_picker.draw(frame, area);
-    }
-}
-
-fn build_port_forward_response(mut input: Vec<String>, resource: ResourceRef) -> ResponseEvent {
-    if input.len() == 3 {
-        let container_port = input[0].parse::<u16>().unwrap_or_default();
-        let local_port = input[1].parse::<u16>().unwrap_or_default();
-        let address = input.pop().unwrap_or_default();
-        ResponseEvent::PortForward(resource, container_port, local_port, address)
-    } else {
-        ResponseEvent::Handled
     }
 }
