+++ b4n/ui/views/shell/kube/view.rs
@@ -112,8 +112,6 @@ impl ShellView {
             return ResponseEvent::Handled;
         }
 
-        let mut size = 22;
-
         let detach = if self.is_attach { "󰕍 back" } else { " detach shell" };
         let mut builder = ActionsListBuilder::default()
             .with_menu_action(ActionItem::menu(2, "󰆒 paste", "paste"))
@@ -128,7 +126,6 @@ impl ShellView {
         }
 
         if self.is_attach && self.bridge.is_application_mode().is_none() {
-            size = 28;
             let mode = if self.is_app_mode { " normal" } else { " app." };
             builder.add_menu_action(ActionItem::menu(3, &format!("{mode} cursor key mode"), "app_mode"));
         }
@@ -141,7 +138,10 @@ impl ShellView {
             }
         }
 
-        self.command_palette = CommandPalette::new(Rc::clone(&self.app_data), builder.build(None), size).to_mouse_menu();
+        let actions = builder.build(None);
+        let width = u16::try_from(actions.max_item_len() + 4).unwrap_or(u16::MAX).max(22);
+
+        self.command_palette = CommandPalette::new(Rc::clone(&self.app_data), actions, width).to_mouse_menu();
         self.command_palette.show_at((x.saturating_sub(3), y).into());
 
         ResponseEvent::Handled
